requirements.txt
praw==7.8.1
python-dotenv
kafka-python
pyspark==3.5.1
azure-storage-file-datalake